Rename setupRequestParsing to setUpRequestParsing for consistency (#755)
